move LoginResults to relevant file
