+++ src/scrapers/base-scraper.ts
@@ -22,13 +22,6 @@ function createGenericError(errorMessage) {
   return createErrorResult(ERRORS.GENERIC, errorMessage);
 }
 
-export enum LoginResults {
-  Success = 'Success',
-  InvalidPassword = 'InvalidPassword',
-  ChangePassword = 'ChangePassword',
-  UnknownError = 'UnknownError',
-}
-
 // TODO es consider browser type
 export interface BaseScraperOptions {
   companyId: string;
